+++ html/fullhouse.py
@@ -21,6 +21,7 @@ print("</head>")
 print("<body>")
 print("<h1>House Energy Usage</h1>")
 print("<a href='comed.py'>Show Only Comed Prices</a><br/>")
+print("<a href='dashboard.html'>New Dashboard (faster)</a><br/>")
 print("<h3>Current time:</h3>%s"%(time.asctime()))
 print('<br/>\n')